Move display and logic to pkg
