+++ pkg/display/display.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/Danice123/idk/display/screen"
-	"github.com/Danice123/idk/display/utils"
+	"github.com/Danice123/idk/pkg/display/screen"
+	"github.com/Danice123/idk/pkg/display/utils"
 	"github.com/faiface/pixel/pixelgl"
 )
