+++ pkg/display/entity/entity.go
@@ -1,8 +1,8 @@
 package entity
 
 import (
-	"github.com/Danice123/idk/display/texturepacker"
-	"github.com/Danice123/idk/logic/entity"
+	"github.com/Danice123/idk/pkg/display/texturepacker"
+	"github.com/Danice123/idk/pkg/logic/entity"
 	"github.com/faiface/pixel"
 	"github.com/faiface/pixel/pixelgl"
 )
