+++ pkg/display/texturepacker/spritesheet.go
@@ -7,7 +7,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/Danice123/idk/display/utils"
+	"github.com/Danice123/idk/pkg/display/utils"
 	"github.com/faiface/pixel"
 )
