+++ pkg/display/utils/utils.go
